+++ scx-dao/src/main/java/cool/scx/dao/dialect/SQLiteDialect.java
@@ -89,7 +89,7 @@ public class SQLiteDialect implements Dialect {
     }
 
     @Override
-    public String getLimitSQL(String sql, Integer offset, Integer rowCount) {
+    public String getLimitSQL(String sql, Long offset, Long rowCount) {
         var limitClauses = rowCount == null ? "" : offset == null || offset == 0 ? " LIMIT " + rowCount : " LIMIT " + offset + "," + rowCount;
         return sql + limitClauses;
     }
